+++ src/main/java/com/manchui/domain/gathering/dto/ClosedGatheringResponse.java
@@ -1,4 +1,4 @@
-package com.manchui.domain.dto.gathering;
+package com.manchui.domain.gathering.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
