+++ src/main/java/com/manchui/domain/gathering/dto/GatheringCreateResponse.java
@@ -1,4 +1,4 @@
-package com.manchui.domain.dto.gathering;
+package com.manchui.domain.gathering.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
