+++ src/main/java/com/manchui/domain/gathering/dto/GatheringPagingResponse.java
@@ -1,4 +1,4 @@
-package com.manchui.domain.dto.gathering;
+package com.manchui.domain.gathering.dto;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
