+++ src/main/java/com/manchui/domain/gathering/entity/Gathering.java
@@ -1,8 +1,8 @@
-package com.manchui.domain.entity;
+package com.manchui.domain.gathering.entity;
 
 import com.manchui.domain.chat.entity.mysql.ChatRoom;
-import com.manchui.domain.dto.gathering.ClosedGatheringInfoResponse;
-import com.manchui.domain.dto.gathering.GatheringCreateResponse;
+import com.manchui.domain.gathering.dto.ClosedGatheringInfoResponse;
+import com.manchui.domain.gathering.dto.GatheringCreateResponse;
 import com.manchui.domain.user.entity.User;
 import com.manchui.global.entity.Timestamped;
 import jakarta.persistence.*;
