+++ src/main/java/com/manchui/domain/gathering/service/GatheringServiceImpl.java
@@ -1,21 +1,29 @@
-package com.manchui.domain.service;
+package com.manchui.domain.gathering.service;
 
+import com.manchui.domain.attendance.entity.Attendance;
+import com.manchui.domain.attendance.repository.AttendanceRepository;
+import com.manchui.domain.auth.dto.CustomUserDetails;
+import com.manchui.domain.chat.dto.ChatMessageRequest;
 import com.manchui.domain.chat.entity.mysql.ChatRoom;
 import com.manchui.domain.chat.entity.mysql.ChatRoomUser;
 import com.manchui.domain.chat.repository.ChatRoomRepository;
 import com.manchui.domain.chat.repository.ChatRoomUserRepository;
 import com.manchui.domain.chat.service.ChatMessageService;
-import com.manchui.domain.auth.dto.CustomUserDetails;
-import com.manchui.domain.user.dto.UserInfo;
-import com.manchui.domain.chat.dto.ChatMessageRequest;
-import com.manchui.domain.dto.gathering.*;
-import com.manchui.domain.dto.review.ReviewDetailPagingResponse;
-import com.manchui.domain.dto.review.ReviewInfo;
-import com.manchui.domain.dto.review.ReviewScoreInfo;
-import com.manchui.domain.entity.*;
+import com.manchui.domain.gathering.dto.*;
+import com.manchui.domain.gathering.entity.Gathering;
+import com.manchui.domain.gathering.repository.GatheringRepository;
+import com.manchui.domain.heart.entity.Heart;
+import com.manchui.domain.heart.repository.HeartRepository;
+import com.manchui.domain.image.entity.Image;
 import com.manchui.domain.notification.entity.NotificationType;
 import com.manchui.domain.notification.service.NotificationServiceImpl;
-import com.manchui.domain.repository.*;
+import com.manchui.domain.image.repository.ImageRepository;
+import com.manchui.domain.review.dto.ReviewDetailPagingResponse;
+import com.manchui.domain.review.dto.ReviewInfo;
+import com.manchui.domain.review.dto.ReviewScoreInfo;
+import com.manchui.domain.review.repository.ReviewRepository;
+import com.manchui.domain.image.service.ImageServiceImpl;
+import com.manchui.domain.user.dto.UserInfo;
 import com.manchui.domain.user.entity.User;
 import com.manchui.domain.user.service.UserService;
 import com.manchui.global.exception.CustomException;
